+++ packages/account-tree-controller/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `selectedAccountGroup` state and bidirectional synchronization with `AccountsController` ([#6186](https://github.com/MetaMask/core/pull/6186))
+  - New `getSelectedAccountGroup()` and `setSelectedAccountGroup()` methods.
+  - Automatic synchronization when selected account changes in AccountsController.
+  - New action types `AccountTreeControllerGetSelectedAccountGroupAction` and `AccountTreeControllerSetSelectedAccountGroupAction`.
+
 ## [0.6.0]
 
 ### Changed